Module google_firebasehosting1_beta1::api

Structs

ActingUser
Contains metadata about the user who performed an action, such as creating a release or finalizing a version.

CertDnsChallenge
Represents a DNS certificate challenge.

CertHttpChallenge
Represents an HTTP certificate challenge.

Channel
A Channel represents a stream of releases for a site. All sites have a default live channel that serves content to the Firebase-provided subdomains and any connected custom domains.

CloneVersionRequest
There is no detailed description.

CloudRunRewrite
A configured rewrite that directs requests to a Cloud Run service. If the Cloud Run service does not exist when setting or updating your Firebase Hosting configuration, then the request fails. Any errors from the Cloud Run service are passed to the end user (for example, if you delete a service, any requests directed to that service receive a 404 error).

Domain
The intended behavior and status information of a domain.

DomainProvisioning
The current certificate provisioning status information for a domain.

DomainRedirect
Defines the behavior of a domain-level redirect. Domain redirects preserve the path of the redirect but replace the requested domain with the one specified in the redirect configuration.

Empty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance: service Foo { rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty); } The JSON representation for Empty is empty JSON object {}.

FirebaseHosting
Central instance to access all FirebaseHosting related resource activities

Header
A Header specifies a URL pattern that, if matched to the request URL path, triggers Hosting to apply the specified custom response headers.

I18nConfig
If provided, i18n rewrites are enabled.

ListChannelsResponse
There is no detailed description.

ListDomainsResponse
The response to listing Domains.

ListReleasesResponse
There is no detailed description.

ListSitesResponse
There is no detailed description.

ListVersionFilesResponse
There is no detailed description.

ListVersionsResponse
There is no detailed description.

Operation
This resource represents a long-running operation that is the result of a network API call.

PathFilter
A representation of filter path.

PopulateVersionFilesRequest
There is no detailed description.

PopulateVersionFilesResponse
There is no detailed description.

PreviewConfig
Deprecated in favor of site channels.

ProjectMethods
A builder providing access to all methods supported on project resources. It is not used directly, but through the FirebaseHosting hub.

ProjectOperationGetCall
Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

ProjectSiteChannelCreateCall
Creates a new channel in the specified site.

ProjectSiteChannelDeleteCall
Deletes the specified channel of the specified site. The live channel cannot be deleted.

ProjectSiteChannelGetCall
Retrieves information for the specified channel of the specified site.

ProjectSiteChannelListCall
Lists the channels for the specified site. All sites have a default live channel.

ProjectSiteChannelPatchCall
Updates information for the specified channel of the specified site. Implicitly creates the channel if it doesn't already exist.

ProjectSiteChannelReleaseCreateCall
Creates a new release, which makes the content of the specified version actively display on the appropriate URL(s).

ProjectSiteChannelReleaseListCall
Lists the releases that have been created for the specified site or channel. When used to list releases for a site, this list includes releases for both the default live channel and any active preview channels for the specified site.

ProjectSiteCreateCall
Creates a new Hosting Site in the specified parent Firebase project. Note that Hosting sites can take several minutes to propagate through Firebase systems.

ProjectSiteDeleteCall
Deletes the specified Hosting Site from the specified parent Firebase project.

ProjectSiteDomainCreateCall
Creates a domain mapping on the specified site.

ProjectSiteDomainDeleteCall
Deletes the existing domain mapping on the specified site.

ProjectSiteDomainGetCall
Gets a domain mapping on the specified site.

ProjectSiteDomainListCall
Lists the domains for the specified site.

ProjectSiteDomainUpdateCall
Updates the specified domain mapping, creating the mapping as if it does not exist.

ProjectSiteGetCall
Gets the specified Hosting Site.

ProjectSiteGetConfigCall
Gets the Hosting metadata for a specific site.

ProjectSiteListCall
Lists each Hosting Site associated with the specified parent Firebase project.

ProjectSitePatchCall
Updates attributes of the specified Hosting Site.

ProjectSiteReleaseCreateCall
Creates a new release, which makes the content of the specified version actively display on the appropriate URL(s).

ProjectSiteReleaseListCall
Lists the releases that have been created for the specified site or channel. When used to list releases for a site, this list includes releases for both the default live channel and any active preview channels for the specified site.

ProjectSiteUpdateConfigCall
Sets the Hosting metadata for a specific site.

ProjectSiteVersionCloneCall
Creates a new version on the specified target site using the content of the specified version.

ProjectSiteVersionCreateCall
Creates a new version for the specified site.

ProjectSiteVersionDeleteCall
Deletes the specified version.

ProjectSiteVersionFileListCall
Lists the remaining files to be uploaded for the specified version.

ProjectSiteVersionListCall
Lists the versions that have been created for the specified site. This list includes versions for both the default live channel and any active preview channels for the specified site.

ProjectSiteVersionPatchCall
Updates the specified metadata for the specified version. This method will fail with FAILED_PRECONDITION in the event of an invalid state transition. The supported state transitions for a version are from CREATED to FINALIZED. Use DeleteVersion to set the status of a version to DELETED.

ProjectSiteVersionPopulateFileCall
Adds content files to the specified version. Each file must be under 2 GB.

Redirect
A Redirect specifies a URL pattern that, if matched to the request URL path, triggers Hosting to respond with a redirect to the specified destination path.

Release
A Release is a particular collection of configurations and files that is set to be public at a particular time.

Rewrite
A Rewrite specifies a URL pattern that, if matched to the request URL path, triggers Hosting to respond as if the service were given the specified destination URL.

ServingConfig
The configuration for how incoming requests to a site should be routed and processed before serving content. The URL request paths are matched against the specified URL patterns in the configuration, then Hosting applies the applicable configuration according to a specific priority order.

Site
A Site represents a Firebase Hosting site.

SiteChannelCreateCall
Creates a new channel in the specified site.

SiteChannelDeleteCall
Deletes the specified channel of the specified site. The live channel cannot be deleted.

SiteChannelGetCall
Retrieves information for the specified channel of the specified site.

SiteChannelListCall
Lists the channels for the specified site. All sites have a default live channel.

SiteChannelPatchCall
Updates information for the specified channel of the specified site. Implicitly creates the channel if it doesn't already exist.

SiteChannelReleaseCreateCall
Creates a new release, which makes the content of the specified version actively display on the appropriate URL(s).

SiteChannelReleaseListCall
Lists the releases that have been created for the specified site or channel. When used to list releases for a site, this list includes releases for both the default live channel and any active preview channels for the specified site.

SiteConfig
A SiteConfig contains metadata associated with a specific site that controls Firebase Hosting serving behavior

SiteDomainCreateCall
Creates a domain mapping on the specified site.

SiteDomainDeleteCall
Deletes the existing domain mapping on the specified site.

SiteDomainGetCall
Gets a domain mapping on the specified site.

SiteDomainListCall
Lists the domains for the specified site.

SiteDomainUpdateCall
Updates the specified domain mapping, creating the mapping as if it does not exist.

SiteGetConfigCall
Gets the Hosting metadata for a specific site.

SiteMethods
A builder providing access to all methods supported on site resources. It is not used directly, but through the FirebaseHosting hub.

SiteReleaseCreateCall
Creates a new release, which makes the content of the specified version actively display on the appropriate URL(s).

SiteReleaseListCall
Lists the releases that have been created for the specified site or channel. When used to list releases for a site, this list includes releases for both the default live channel and any active preview channels for the specified site.

SiteUpdateConfigCall
Sets the Hosting metadata for a specific site.

SiteVersionCloneCall
Creates a new version on the specified target site using the content of the specified version.

SiteVersionCreateCall
Creates a new version for the specified site.

SiteVersionDeleteCall
Deletes the specified version.

SiteVersionFileListCall
Lists the remaining files to be uploaded for the specified version.

SiteVersionListCall
Lists the versions that have been created for the specified site. This list includes versions for both the default live channel and any active preview channels for the specified site.

SiteVersionPatchCall
Updates the specified metadata for the specified version. This method will fail with FAILED_PRECONDITION in the event of an invalid state transition. The supported state transitions for a version are from CREATED to FINALIZED. Use DeleteVersion to set the status of a version to DELETED.

SiteVersionPopulateFileCall
Adds content files to the specified version. Each file must be under 2 GB.

Status
The Status type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by gRPC. Each Status message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the API Design Guide.

Version
A Version is a configuration and a collection of static files which determine how a site is displayed.

VersionFile
A static content file that is part of a version.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.