Struct google_firebasehosting1_beta1::SiteMethods

pub struct SiteMethods<'a, C, A> where
    C: 'a,
    A: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on site resources. It is not used directly, but through the FirebaseHosting hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_firebasehosting1_beta1 as firebasehosting1_beta1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use firebasehosting1_beta1::FirebaseHosting;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())),
                              <MemoryStorage as Default>::default(), None);
let mut hub = FirebaseHosting::new(hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `domains_create(...)`, `domains_delete(...)`, `domains_get(...)`, `domains_list(...)`, `domains_update(...)`, `get_config(...)`, `releases_create(...)`, `releases_list(...)`, `update_config(...)`, `versions_create(...)`, `versions_delete(...)`, `versions_files_list(...)`, `versions_patch(...)` and `versions_populate_files(...)`
// to build up your call.
let rb = hub.sites();

Methods

impl<'a, C, A> SiteMethods<'a, C, A>

pub fn versions_files_list(
&self,
parent: &str
) -> SiteVersionFileListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists the remaining files to be uploaded for the specified version.

Arguments

parent - Required. The parent to list files for, in the format: sites/site-name/versions/versionID

pub fn domains_delete(&self, name: &str) -> SiteDomainDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes the existing domain mapping on the specified site.

Arguments

name - Required. The name of the domain association to delete.

pub fn releases_list(&self, parent: &str) -> SiteReleaseListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists the releases that have been created on the specified site.

Arguments

parent - Required. The parent for which to list files, in the format: sites/site-name

pub fn releases_create(
&self,
request: Release,
parent: &str
) -> SiteReleaseCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a new release which makes the content of the specified version actively display on the site.

Arguments

request - No description provided.

parent - The site that the release belongs to, in the format: sites/site-name

pub fn domains_create(
&self,
request: Domain,
parent: &str
) -> SiteDomainCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a domain mapping on the specified site.

Arguments

request - No description provided.

parent - Required. The parent to create the domain association for, in the format: sites/site-name

pub fn domains_list(&self, parent: &str) -> SiteDomainListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists the domains for the specified site.

Arguments

parent - Required. The parent for which to list domains, in the format: sites/site-name

pub fn versions_create(
&self,
request: Version,
parent: &str
) -> SiteVersionCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a new version for a site.

Arguments

request - No description provided.

parent - Required. The parent to create the version for, in the format: sites/site-name

pub fn versions_patch(
&self,
request: Version,
name: &str
) -> SiteVersionPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates the specified metadata for a version. Note that this method will fail with FAILED_PRECONDITION in the event of an invalid state transition. The only valid transition for a version is currently from a CREATED status to a FINALIZED status. Use DeleteVersion to set the status of a version to DELETED.

Arguments

request - No description provided.

name - The unique identifier for a version, in the format: sites/site-name/versions/versionID This name is provided in the response body when you call the CreateVersion endpoint.

pub fn update_config(
&self,
request: SiteConfig,
name: &str
) -> SiteUpdateConfigCall<'a, C, A>

Create a builder to help you perform the following task:

Sets the Hosting metadata for a specific site.

Arguments

request - No description provided.

name - Required. The site for which to update the SiteConfig, in the format: sites/site-name/config

pub fn versions_populate_files(
&self,
request: PopulateVersionFilesRequest,
parent: &str
) -> SiteVersionPopulateFileCall<'a, C, A>

Create a builder to help you perform the following task:

Adds content files to a version.

Arguments

request - No description provided.

parent - Required. The version to add files to, in the format: sites/site-name/versions/versionID

pub fn get_config(&self, name: &str) -> SiteGetConfigCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the Hosting metadata for a specific site.

Arguments

name - Required. The site for which to get the SiteConfig, in the format: sites/site-name/config

pub fn domains_get(&self, name: &str) -> SiteDomainGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets a domain mapping on the specified site.

Arguments

name - Required. The name of the domain configuration to get.

pub fn versions_delete(&self, name: &str) -> SiteVersionDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes the specified version.

Arguments

name - Required. The name of the version to be deleted, in the format: sites/site-name/versions/versionID

pub fn domains_update(
&self,
request: Domain,
name: &str
) -> SiteDomainUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates the specified domain mapping, creating the mapping as if it does not exist.

Arguments

request - No description provided.

name - Required. The name of the domain association to update or create, if an association doesn't already exist.